refactor: 토큰 관련 애노테이션 디렉토리 이동
