+++ src/main/java/com/kgu/studywithme/auth/controller/OAuthApiController.java
@@ -3,7 +3,7 @@ package com.kgu.studywithme.auth.controller;
 import com.kgu.studywithme.auth.infra.oauth.OAuthUri;
 import com.kgu.studywithme.auth.service.OAuthService;
 import com.kgu.studywithme.auth.service.dto.response.LoginResponse;
-import com.kgu.studywithme.global.annotation.ExtractPayload;
+import com.kgu.studywithme.auth.utils.ExtractPayload;
 import com.kgu.studywithme.global.dto.SimpleResponseWrapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
