+++ src/main/java/com/kgu/studywithme/auth/utils/ExtractPayloadArgumentResolver.java
@@ -1,8 +1,6 @@
-package com.kgu.studywithme.global.annotation;
+package com.kgu.studywithme.auth.utils;
 
 import com.kgu.studywithme.auth.exception.AuthErrorCode;
-import com.kgu.studywithme.auth.utils.AuthorizationExtractor;
-import com.kgu.studywithme.auth.utils.JwtTokenProvider;
 import com.kgu.studywithme.global.exception.StudyWithMeException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.MethodParameter;
