+++ src/main/java/com/kgu/studywithme/study/controller/StudyApiController.java
@@ -1,7 +1,7 @@
 package com.kgu.studywithme.study.controller;
 
+import com.kgu.studywithme.auth.utils.ExtractPayload;
 import com.kgu.studywithme.global.annotation.CheckStudyHost;
-import com.kgu.studywithme.global.annotation.ExtractPayload;
 import com.kgu.studywithme.study.controller.dto.request.StudyRegisterRequest;
 import com.kgu.studywithme.study.controller.dto.request.StudyUpdateRequest;
 import com.kgu.studywithme.study.service.StudyService;
