+++ src/main/java/com/kgu/studywithme/study/controller/StudyReviewApiController.java
@@ -1,6 +1,6 @@
 package com.kgu.studywithme.study.controller;
 
-import com.kgu.studywithme.global.annotation.ExtractPayload;
+import com.kgu.studywithme.auth.utils.ExtractPayload;
 import com.kgu.studywithme.study.controller.dto.request.ReviewRequest;
 import com.kgu.studywithme.study.service.StudyReviewService;
 import lombok.RequiredArgsConstructor;
