+++ src/main/java/com/kgu/studywithme/study/controller/attendance/AttendanceApiController.java
@@ -1,7 +1,7 @@
 package com.kgu.studywithme.study.controller.attendance;
 
+import com.kgu.studywithme.auth.utils.ExtractPayload;
 import com.kgu.studywithme.global.annotation.CheckStudyHost;
-import com.kgu.studywithme.global.annotation.ExtractPayload;
 import com.kgu.studywithme.study.controller.dto.request.AttendanceRequest;
 import com.kgu.studywithme.study.service.attendance.AttendanceService;
 import lombok.RequiredArgsConstructor;
